reorganized file system, removed cryptocurrencies from list
